+++ tests/examples/opinions/united_states/colo_example.compare.json
@@ -1,74 +1,26 @@
 [
   {
-    "case_dates": "2023-11-14",
-    "case_names": "in Re People in the Interest of Minor Children J.P.",
-    "download_urls": "https://www.courts.state.co.us//userfiles/file/Court_Probation/Supreme_Court/Opinions/2023/23SA126.pdf",
+    "case_dates": "2024-07-01",
+    "case_names": "Ricardo Castro v. The People of the State of Colorado",
+    "download_urls": "https://colorado.vlex.io/vid/castro-v-people-1042407550/content",
     "precedential_statuses": "Published",
     "blocked_statuses": false,
     "date_filed_is_approximate": false,
-    "docket_numbers": "23SA126",
-    "judges": "Justin Haenlein",
-    "summaries": "The petitioner seeks relief from the trial court's order of April 7, 2023. On May 18, 2023, the Supreme Court issued a rule to show cause why the trial court did not err in ordering the petitioner to turn over potentially privileged documents for in camera review. The respondents are directed to file a written answer on or before June 8, 2023. The petitioner has 14 days from receipt of the answer to reply. Opinion issued November 14, 2023",
+    "docket_numbers": "22SC712",
+    "citations": "2024 CO 56",
+    "parallel_citations": "",
     "case_name_shorts": ""
   },
   {
-    "case_dates": "2023-10-23",
-    "case_names": "In Re: People v. Walthour, Ashleigh",
-    "download_urls": "https://www.courts.state.co.us//userfiles/file/Court_Probation/Supreme_Court/Opinions/2023/23SA125.pdf",
-    "precedential_statuses": "Published",
+    "case_dates": "2024-07-01",
+    "case_names": "In re the Marriage of Elyssa M. Fox, and Alexander L. Speaker",
+    "download_urls": "https://colorado.vlex.io/vid/in-re-marriage-of-1042461692/content",
+    "precedential_statuses": "Unknown",
     "blocked_statuses": false,
     "date_filed_is_approximate": false,
-    "docket_numbers": "23SA125",
-    "judges": "Joshua Williford",
-    "summaries": "The petitioner seeks relief from the trial court's order of March 7, 2023. On May 12, 2023, the Supreme Court issued a rule to show cause why the trial court did not err in suppressing the results of an impending blood test because it hadn't been completed by the court's prior deadline. The respondents are directed to file a written answer on or before June 9, 2023. The petitioner has 21 days from receipt of the answer to reply. Opinion issued October 23, 2023",
-    "case_name_shorts": ""
-  },
-  {
-    "case_dates": "2023-10-16",
-    "case_names": "In Re People v. Seymour, Gavin",
-    "download_urls": "https://www.courts.state.co.us//userfiles/file/Court_Probation/Supreme_Court/Opinions/2023/23SA12.pdf",
-    "precedential_statuses": "Published",
-    "blocked_statuses": false,
-    "date_filed_is_approximate": false,
-    "docket_numbers": "23SA12",
-    "judges": "Martin Egelhoff",
-    "summaries": "The petitioner seeks relief from the trial court's Order of November 16, 2022. On January 17, 2023, the Supreme Court issued a rule to show cause why the trial court did not err in denying the defendant's motion to suppress. The respondents are directed to file a written Answer on or before February 14, 2023. The petitioner has 21 days from receipt of the Answer within which to Reply. Opinion issued October 16, 2023",
-    "case_name_shorts": ""
-  },
-  {
-    "case_dates": "2023-09-25",
-    "case_names": "In Re Edwards, Tana v. New Century Hospice",
-    "download_urls": "https://www.courts.state.co.us//userfiles/file/Court_Probation/Supreme_Court/Opinions/2023/23SA91.pdf",
-    "precedential_statuses": "Published",
-    "blocked_statuses": false,
-    "date_filed_is_approximate": false,
-    "docket_numbers": "23SA91",
-    "judges": "Mark Bailey",
-    "summaries": "The petitioners seek relief from the trial court's order of December 19, 2022. On April 3, 2023, the Supreme Court issued a rule to show cause why the trial court did not err in denying the defendants' motion for summary judgment on the plaintiff's claim for negligent supervision. The respondents are directed to file a written answer on or before May 1, 2023. The petitioners have 21 days from receipt of the answer within which to reply. Opinion issued September 25, 2023",
-    "case_name_shorts": ""
-  },
-  {
-    "case_dates": "2023-06-20",
-    "case_names": "In Re Smith, Jerrelle v. People",
-    "download_urls": "https://www.courts.state.co.us//userfiles/file/Court_Probation/Supreme_Court/Opinions/2023/23SA2.pdf",
-    "precedential_statuses": "Published",
-    "blocked_statuses": false,
-    "date_filed_is_approximate": false,
-    "docket_numbers": "23SA2",
-    "judges": "Robert Kiesnowski",
-    "summaries": "On January 6, 2023, the Supreme Court issued a rule to show cause why the trial court did not err in denying bond to the defendant. The respondents are directed to file a written answer on or before February 3, 2023. The petitioner has 21 days from receipt of the answer within which to reply. Opinion issued 6/20/23",
-    "case_name_shorts": ""
-  },
-  {
-    "case_dates": "2023-06-05",
-    "case_names": "In Re People v. Kelley, Noelle",
-    "download_urls": "https://www.courts.state.co.us//userfiles/file/Court_Probation/Supreme_Court/Opinions/2022/22SA874.pdf",
-    "precedential_statuses": "Published",
-    "blocked_statuses": false,
-    "date_filed_is_approximate": false,
-    "docket_numbers": "22SA874",
-    "judges": "Eric White",
-    "summaries": "The petitioner seeks relief from the trial court's order of November 23, 2022. On December 5, 2022, the Supreme Court issued a rule to show cause why the trial court did not err in (1) permitting the prosecution to present evidence that the defendant refused to waive her privilege, and (2) ruling that the defendant waived her privilege. The respondents are directed to file a written answer on or before January 3, 2023. The petitioner has 21 days from receipt of the answer within which to reply.t Opinion issued on June 5, 2023",
+    "docket_numbers": "24SC76",
+    "citations": "",
+    "parallel_citations": "",
     "case_name_shorts": ""
   }
 ]
