+++ tests/examples/opinions/united_states/coloctapp_example.compare.json
@@ -1,22 +1,26 @@
 [
   {
-    "case_dates": "2023-11-16",
-    "case_names": "Samuel Perez v. By the Rockies, LLC",
-    "download_urls": "https://www.courts.state.co.us/Courts/Court_of_Appeals/Opinion/2023/22CA1791-PD.pdf",
-    "precedential_statuses": "Published",
+    "case_dates": "2024-05-16",
+    "case_names": "The People of the State of Colorado, In the Interest of T.C., Jr., a Child, and Concerning L.N.P. and T.C.",
+    "download_urls": "https://colorado.vlex.io/vid/people-ex-rel-t-1035297640/content",
+    "precedential_statuses": "Unknown",
     "blocked_statuses": false,
     "date_filed_is_approximate": false,
-    "docket_numbers": "22CA1791",
+    "docket_numbers": "23CA1539",
+    "citations": "",
+    "parallel_citations": "",
     "case_name_shorts": ""
   },
   {
-    "case_dates": "2023-11-16",
-    "case_names": "Chance Gresser v. Banner Health",
-    "download_urls": "https://www.courts.state.co.us/Courts/Court_of_Appeals/Opinion/2023/22CA1502-PD.pdf",
+    "case_dates": "2024-05-16",
+    "case_names": "Andrew Ortiz v. Progressive Direct Insurance Company",
+    "download_urls": "https://colorado.vlex.io/vid/ortiz-v-progressive-direct-1035272272/content",
     "precedential_statuses": "Published",
     "blocked_statuses": false,
     "date_filed_is_approximate": false,
-    "docket_numbers": "22CA1502",
+    "docket_numbers": "23CA0292",
+    "citations": "2024 COA 54",
+    "parallel_citations": "",
     "case_name_shorts": ""
   }
 ]
